airflow/requirements.txt
apache-airflow-providers-google
requests
pandas
google-cloud-bigquery
google-cloud-storage
python-dotenv
